add timer compo and readonly to subject in listener
