+++ orders/src/events/listeners/expiration-complete-listener.ts
@@ -10,7 +10,7 @@ import { Order } from "../../models/order";
 import { OrderCancelledPublisher } from "../publishers/order-cancelled-publisher";
 
 export class ExpirationCompleteListener extends Listener<ExpirationCompleteEvent> {
-  subject: Subjects.ExpirationComplete = Subjects.ExpirationComplete;
+  readonly subject = Subjects.ExpirationComplete;
   queueGroupName = queueGroupName;
 
   async onMessage(data: ExpirationCompleteEvent["data"], msg: Message) {
