+++ orders/src/events/listeners/payment-created-listener.ts
@@ -9,7 +9,7 @@ import { queueGroupName } from "../queue-group-name";
 import { Order } from "../../models/order";
 
 export class PaymentCreatedListener extends Listener<PaymentCreatedEvent> {
-  subject: Subjects.PaymentCreated = Subjects.PaymentCreated;
+  readonly subject = Subjects.PaymentCreated;
   queueGroupName = queueGroupName;
 
   async onMessage(data: PaymentCreatedEvent["data"], msg: Message) {
