+++ orders/src/events/listeners/ticket-created-listener.ts
@@ -4,7 +4,7 @@ import { Ticket } from "../../models/ticket";
 import { queueGroupName } from "../queue-group-name";
 
 export class TicketCreatedListener extends Listener<TicketCreatedEvent> {
-  subject: Subjects.TicketCreated = Subjects.TicketCreated;
+  readonly subject = Subjects.TicketCreated;
   queueGroupName = queueGroupName;
 
   async onMessage(data: TicketCreatedEvent["data"], msg: Message) {
