+++ orders/src/routes/new.ts
@@ -15,7 +15,7 @@ import { natsWrapper } from "../nats-wrapper";
 
 const router = express.Router();
 
-const EXPIRATION_WINDOW_SECONDS = 2 * 60;
+const EXPIRATION_WINDOW_SECONDS = 0.2 * 60;
 
 router.post(
   "/api/orders",
