+++ payments/src/events/listeners/order-created-listener.ts
@@ -4,7 +4,7 @@ import { queueGroupName } from "../queue-group-name";
 import { Order } from "../../models/order";
 
 export class OrderCreatedListener extends Listener<OrderCreatedEvent> {
-  subject: Subjects.OrderCreated = Subjects.OrderCreated;
+  readonly subject = Subjects.OrderCreated;
   queueGroupName = queueGroupName;
 
   async onMessage(data: OrderCreatedEvent["data"], msg: Message) {
